Adicionando uma nova classe aluguel e colocando um novo iframe
