+++ web/index.jsp
@@ -13,25 +13,32 @@
     </head>
     <body>
         <div>
-            <form action="Aluga.jsp" method="get" target="result">
+            <form action="aluga.jsp" method="get" target="result2">
                 <h1>Aluguel de carros!</h1>
-                <p>Bem vindo a um teste de aluguel de carros, para alugar um automóvel digite o seu <strong>nome</strong><br> 
+                <p>Bem vindo a um teste de aluguel de carros, para alugar um 
+                    automóvel digite o seu <strong>nome</strong><br> 
                     e quantas <strong>horas</strong> deseja alugar um carro.</p>
                 <br>
-                Nome:<br><input type="text" name="nome" placeholder="Digite o seu nome"><br>
+                Nome:<br><input type="text" name="nome" 
+                                placeholder="Digite o seu nome"><br>
                 <br>
-                Hora:<br><input type="number" name="hora" placeholder="Digite quantas horas">
+                Hora:<br><input type="number" name="hora" 
+                                placeholder="Digite quantas horas">
                 <br><br>
-                Modelo do Carro:<br><input type="text" name="modelo" placeholder="Digite o modelo do carro"> 
+                Modelo do Carro:<br><input type="text" name="modelo" 
+                                           placeholder="Digite o modelo do carro"> 
                 
                 <input type="submit" name="enviar" value="Buscar">
-           </form>             
+            </form> 
+           
         </div>
         <br>
         <div id="retorno">
-            <iframe name="result" width="300" height="100">
+            <iframe name="result" src="lista.jsp" width="300" height="100">
                 
             </iframe>
+            
+            <iframe name="result2" src="aluga.jsp" width="300" height="100"></iframe>
             
         </div>    
     </body>
